Create ProjectTasksAdapter.kt

+++ app/src/main/java/com/ahinfo/ahteam/data/parser/detailsProject/remote/dto/ResponseGetProjectTasks.kt
@@ -29,7 +29,7 @@ data class ElementsItem(
 	val name: String? = null,
 
 	@field:SerializedName("unix_time")
-	val unixTime: Int? = null,
+	val unixTime: Long? = null,
 
 	@field:SerializedName("description")
 	val description: String? = null,
